feat(editor): implement view extension manager with builder pattern (#12193)
